requirements.txt
asgiref==3.8.1
Django==4.2.13
djangorestframework==3.15.2
djongo==1.3.6
dnspython==2.7.0
pymongo==3.12.3
pytz==2024.2
sqlparse==0.4.4
tzdata==2024.2
